consolidate preview tests
